+++ meerschaum/plugins/__init__.py
@@ -12,8 +12,10 @@ from meerschaum.utils.threading import Lock, RLock
 from meerschaum.plugins._Plugin import Plugin
 
 _api_plugins: Dict[str, List[Callable[['fastapi.App'], Any]]] = {}
+_sync_hooks: Dict[str, List[Callable[[Any], Any]]] = {}
 _locks = {
     '_api_plugins': RLock(),
+    '_sync_hooks': RLock(),
     '__path__': RLock(),
     'sys.path': RLock(),
     'internal_plugins': RLock(),
@@ -23,6 +25,7 @@ _locks = {
 __all__ = (
     "Plugin", "make_action", "api_plugin", "import_plugins",
     "reload_plugins", "get_plugins", "get_data_plugins", "add_plugin_argument",
+    "sync_hook",
 )
 __pdoc__ = {
     'venvs': False, 'data': False, 'stack': False, 'plugins': False,
@@ -81,6 +84,41 @@ def make_action(
     return function
 
 
+def sync_hook(
+        function: Callable[[Any], Any],
+    ) -> Callable[[Any], Any]:
+    """
+    Register a function as a sync hook to be executed upon completion of a sync.
+    
+    Parameters
+    ----------
+    function: Callable[[Any], Any]
+        The function to execute upon completion of a sync.
+        
+    Returns
+    -------
+    Another function (this is a decorator function).
+
+    Examples
+    --------
+    >>> from meerschaum.plugins import sync_hook
+    >>>
+    >>> @sync_hook
+    ... def log_sync(pipe, success_tuple, duration=None, **kwargs):
+    ...     print(f"It took {round(duration, 2)} seconds to sync {pipe}.")
+    >>>
+    """
+    with _locks['_sync_hooks']:
+        try:
+            if function.__module__ not in _sync_hooks:
+                _sync_hooks[function.__module__] = []
+            _sync_hooks[function.__module__].append(function)
+        except Exception as e:
+            from meerschaum.utils.warnings import warn
+            warn(e)
+    return function
+
+
 def api_plugin(function: Callable[[Any], Any]) -> Callable[[Any], Any]:
     """
     Execute the function when initializing the Meerschaum API module.
@@ -106,7 +144,7 @@ def api_plugin(function: Callable[[Any], Any]) -> Callable[[Any], Any]:
     >>> def initialize_plugin(app):
     ...     @app.get('/my/new/path')
     ...     def new_path():
-    ...         return {'message' : 'It works!'}
+    ...         return {'message': 'It works!'}
     >>>
     """
     with _locks['_api_plugins']: