+++ docs/mkdocs/news/changelog.md
@@ -4,6 +4,21 @@
 
 This is the current release cycle, so stay tuned for future releases!
 
+### v2.1.3
+
+- **Add the decorator `@sync_hook`.**  
+  The new decorator `@sync_hook` lets you intercept a Pipe and its success tuple on every sync.
+
+  ```python
+  import meerschaum as mrsm
+  from meerschaum.plugins import sync_hook
+
+  @sync_hook
+  def log_sync(pipe, success_tuple, duration=0):
+      print(f"It took {round(duration, 2)} seconds to sync {pipe}.")
+      mrsm.pprint(success_tuple)
+  ```
+
 ### v2.1.1 – v2.1.2
 
 - **Add `upsert` for high-performance pipes.**